refactor: separate line modal component(create/edit)
